January 11, 2022 3:43 PM

+++ cutout_tester.py
@@ -8,7 +8,7 @@ import numpy as np
 
 #load image
 img = Image.open('/content/Sketch-Simulator/test_images/eedb70bc-7a45-41cd-98e1-1f91f6285803.jpeg')
-img = torch.from_numpy(np.array(img)).int().unsqueeze(0).permute(0,3,1,2)
+img = torch.from_numpy(np.array(img)).int().unsqueeze(0).permute(0,3,1,2).to(torch.float32)
 print(img.shape)
 
 cutter(img)
